pyproject.toml
[project]
name = "papertrail"
version = "0.1.0"
description = "AI-powered research paper discovery, analysis, and synthesis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "pydantic>=2.0.0",
    "pymupdf>=1.24.0",
    "requests>=2.32.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]

[project.scripts]
papertrail = "papertrail.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
